requirements.txt
# Only add explicit dependencies with dep==version
essentia-tensorflow==2.1b6.dev374
numpy==1.20.1
